add request body validator

+++ src/routes/offers.js
@@ -2,6 +2,13 @@ import express, { query } from 'express';
 import db from '../db/conn.js';
 import isRequired from '../utils/auth-utils.js';
 import { ObjectId } from 'mongodb';
+import {
+  validateResult,
+  createOfferValidator,
+  getOfferByIdValidator,
+  deleteOfferValidator,
+  updateOfferValidator,
+} from '../validators/offers.validator.js';
 
 const router = express.Router();
 
@@ -24,17 +31,27 @@ router.get('/', isRequired, async (req, res) => {
 });
 
 /* This is a get request that is looking for a specific offer. */
-router.get('/idOffer/:idOffer', isRequired, async (req, res) => {
-  let collection = db.collection('offers');
-  let result = await collection.findOne({
-    _id: new ObjectId(req.params.idOffer),
-  });
-  if (result?.userId === res.locals.userId) {
-    res.send(result).status(200);
-  } else {
-    res.sendStatus(404);
+router.get(
+  '/idOffer/:idOffer',
+  getOfferByIdValidator,
+  isRequired,
+  async (req, res) => {
+    const validator = validateResult(req, res);
+    console.log(validator);
+    if (validator.length) {
+      return res.send(validator).status(400);
+    }
+    let collection = db.collection('offers');
+    let result = await collection.findOne({
+      _id: new ObjectId(req.params.idOffer),
+    });
+    if (result?.userId === res.locals.userId) {
+      res.send(result).status(200);
+    } else {
+      res.sendStatus(404);
+    }
   }
-});
+);
 
 /* This is a get request that is looking user offers. */
 router.get('/user', isRequired, async (req, res) => {
@@ -44,7 +61,11 @@ router.get('/user', isRequired, async (req, res) => {
 });
 
 /* Creating a new document in the database. */
-router.post('/', isRequired, async (req, res) => {
+router.post('/', createOfferValidator, isRequired, async (req, res) => {
+  const validator = validateResult(req, res);
+  if (validator.length) {
+    return res.send(validator).status(400);
+  }
   let collection = db.collection('offers');
   let newDocument = req.body;
   newDocument.date = new Date();
@@ -54,37 +75,55 @@ router.post('/', isRequired, async (req, res) => {
 });
 
 /* Deleting an entry from the database. */
-router.delete('/idOffer/:idOffer', isRequired, async (req, res) => {
-  const query = { _id: req.params.idOffer };
-  const collection = db.collection('offers');
-  const offer = await collection.findOne({ _id: new ObjectId(query._id) });
-  if (offer?.userId === res.locals.userId) {
-    let result = await collection.deleteOne({
-      _id: new ObjectId(query._id),
-    });
-    res.send(result).status(200);
-  } else {
-    res.sendStatus(404);
+router.delete(
+  '/idOffer/:idOffer',
+  deleteOfferValidator,
+  isRequired,
+  async (req, res) => {
+    const validator = validateResult(req, res);
+    if (validator.length) {
+      return res.send(validator).status(400);
+    }
+    const query = { _id: req.params.idOffer };
+    const collection = db.collection('offers');
+    const offer = await collection.findOne({ _id: new ObjectId(query._id) });
+    if (offer?.userId === res.locals.userId) {
+      let result = await collection.deleteOne({
+        _id: new ObjectId(query._id),
+      });
+      res.send(result).status(200);
+    } else {
+      res.sendStatus(404);
+    }
   }
-});
+);
 
 /* Updating an offer document from the database */
-router.put('/idOffer/:idOffer', isRequired, async (req, res) => {
-  const filter = { _id: new ObjectId(req.params.idOffer) };
-  const collection = db.collection('offers');
-  const offer = await collection.findOne(filter);
-  if (offer?.userId === res.locals.userId) {
-    const updateDoc = {
-      $set: {
-        isActive: !offer.isActive,
-      },
-    };
-    const options = { upsert: false };
-    const result = await collection.updateOne(filter, updateDoc, options);
-    res.send(result).status(200);
-  } else {
-    res.sendStatus(404);
+router.put(
+  '/idOffer/:idOffer',
+  updateOfferValidator,
+  isRequired,
+  async (req, res) => {
+    const validator = validateResult(req, res);
+    if (validator.length) {
+      return res.send(validator).status(400);
+    }
+    const filter = { _id: new ObjectId(req.params.idOffer) };
+    const collection = db.collection('offers');
+    const offer = await collection.findOne(filter);
+    if (offer?.userId === res.locals.userId) {
+      const updateDoc = {
+        $set: {
+          isActive: !offer.isActive,
+        },
+      };
+      const options = { upsert: false };
+      const result = await collection.updateOne(filter, updateDoc, options);
+      res.send(result).status(200);
+    } else {
+      res.sendStatus(404);
+    }
   }
-});
+);
 
 export default router;
